requirements.txt
streamlit==1.33.0
plotly>=5.20

pandas>=2.2
numpy>=1.24
scikit-learn>=1.4
statsmodels>=0.14
pmdarima==2.0.4
prophet==1.1.5
tensorflow-cpu==2.16.1
joblib>=1.3
backtrader==1.9.78.123